pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "simforge"
version = "0.1.0"
description = "Genesis-powered robot simulator with joint and Cartesian control"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Simforge Team"}]
license = {text = "Proprietary"}
dependencies = [
  "genesis-world>=0.3.3",
  "numpy>=1.24",
  "PyYAML>=6.0",
  "xxhash"
]

[project.urls]
Homepage = "https://example.com/simforge"

[project.scripts]
simforge = "simforge:main"

[project.optional-dependencies]
planning = ["ompl>=1.6"]
qt = ["PySide6>=6.6"]
dev = ["pytest>=7.0", "ruff>=0.5.0"]

[tool.setuptools.packages.find]
include = ["simforge*"]
exclude = ["assets*", "env_configs*"]
